+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added Start Time to *Animate Data* nodes.
 - Added Plane-Axis for *Circle* mode in *Distribute Matrices* node.
 - Added *Material By Name* node.
+- Added *Apply Modifiers* input to the *Splines From Object* node.
 
 ### Fixed
 
@@ -35,6 +36,7 @@
 - Removed *Update Object Matrices* node.
 - The output of the *Get Selected Objects* node is no longer ordered.
 - Selection order now need to be recorded for *Integer ID Keys*.
+- Splines can now be computed from Text objects.
 
 ## 2.2.1 (13 January 2021)